requirements.txt
openai
python-dotenv
scikit-learn
pandas
matplotlib
seaborn
